cpp: Fix argument type in dummy, #3492.
git-svn-id: http://svn.code.sourceforge.net/p/sdcc/code/trunk/sdcc@13738 4a8a32a2-be11-0410-ad9d-d568d2c75423

+++ support/cpp/gcc/cc1_dummies.cc
@@ -162,7 +162,7 @@ bool lang_GNU_C()
 {
 	return true;
 }
-void lhd_append_data(void const*, unsigned long, void*)
+void lhd_append_data(void const*, size_t, void*)
 { SDCPP_DUMMY_FCT();
 }
 void lhd_begin_section(char const*)
